# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  wishlist: 'wishlist',
  applied: 'applied',
  oa_received: 'oa_received',
  interview_scheduled: 'interview_scheduled',
  final_round: 'final_round',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  userId: string;
  company: string;
  role: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  salary?: string | null;
  appliedDate: string;
  status: ApplicationStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  jobUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ApplicationInputStatus = typeof ApplicationInputStatus[keyof typeof ApplicationInputStatus];


export const ApplicationInputStatus = {
  wishlist: 'wishlist',
  applied: 'applied',
  oa_received: 'oa_received',
  interview_scheduled: 'interview_scheduled',
  final_round: 'final_round',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface ApplicationInput {
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  role: string;
  location?: string;
  salary?: string;
  appliedDate: string;
  status: ApplicationInputStatus;
  notes?: string;
  jobUrl?: string;
}

export type ApplicationUpdateStatus = typeof ApplicationUpdateStatus[keyof typeof ApplicationUpdateStatus];


export const ApplicationUpdateStatus = {
  wishlist: 'wishlist',
  applied: 'applied',
  oa_received: 'oa_received',
  interview_scheduled: 'interview_scheduled',
  final_round: 'final_round',
  offer: 'offer',
  rejected: 'rejected',
} as const;

export interface ApplicationUpdate {
  /** @minLength 1 */
  company?: string;
  /** @minLength 1 */
  role?: string;
  location?: string;
  salary?: string;
  appliedDate?: string;
  status?: ApplicationUpdateStatus;
  notes?: string;
  jobUrl?: string;
}

export interface Recruiter {
  id: number;
  userId: string;
  name: string;
  company: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  followUpDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RecruiterInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  company: string;
  email?: string;
  linkedin?: string;
  phone?: string;
  notes?: string;
  followUpDate?: string;
}

export interface RecruiterUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  company?: string;
  email?: string;
  linkedin?: string;
  phone?: string;
  notes?: string;
  followUpDate?: string;
}

export interface Resume {
  id: number;
  userId: string;
  name: string;
  category: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  isDefault?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ResumeInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  content?: string;
  fileUrl?: string;
  isDefault?: boolean;
}

export interface ResumeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  category?: string;
  content?: string;
  fileUrl?: string;
  isDefault?: boolean;
}

export type InterviewQuestionCategory = typeof InterviewQuestionCategory[keyof typeof InterviewQuestionCategory];


export const InterviewQuestionCategory = {
  behavioral: 'behavioral',
  technical: 'technical',
  company_specific: 'company_specific',
  general: 'general',
} as const;

export type InterviewQuestionDifficulty = typeof InterviewQuestionDifficulty[keyof typeof InterviewQuestionDifficulty];


export const InterviewQuestionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface InterviewQuestion {
  id: number;
  userId: string;
  question: string;
  /** @nullable */
  answer?: string | null;
  category: InterviewQuestionCategory;
  difficulty?: InterviewQuestionDifficulty;
  createdAt: string;
  updatedAt: string;
}

export type InterviewQuestionInputCategory = typeof InterviewQuestionInputCategory[keyof typeof InterviewQuestionInputCategory];


export const InterviewQuestionInputCategory = {
  behavioral: 'behavioral',
  technical: 'technical',
  company_specific: 'company_specific',
  general: 'general',
} as const;

export type InterviewQuestionInputDifficulty = typeof InterviewQuestionInputDifficulty[keyof typeof InterviewQuestionInputDifficulty];


export const InterviewQuestionInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface InterviewQuestionInput {
  /** @minLength 1 */
  question: string;
  answer?: string;
  category: InterviewQuestionInputCategory;
  difficulty?: InterviewQuestionInputDifficulty;
}

export type InterviewQuestionUpdateCategory = typeof InterviewQuestionUpdateCategory[keyof typeof InterviewQuestionUpdateCategory];


export const InterviewQuestionUpdateCategory = {
  behavioral: 'behavioral',
  technical: 'technical',
  company_specific: 'company_specific',
  general: 'general',
} as const;

export type InterviewQuestionUpdateDifficulty = typeof InterviewQuestionUpdateDifficulty[keyof typeof InterviewQuestionUpdateDifficulty];


export const InterviewQuestionUpdateDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface InterviewQuestionUpdate {
  /** @minLength 1 */
  question?: string;
  answer?: string;
  category?: InterviewQuestionUpdateCategory;
  difficulty?: InterviewQuestionUpdateDifficulty;
}

export interface ResumeAnalysisInput {
  resumeText: string;
  jobDescription: string;
}

export interface ResumeAnalysisResult {
  atsScore: number;
  feedback: string;
  missingKeywords: string[];
  skillGaps: string[];
  suggestions: string[];
}

export interface CoverLetterInput {
  jobDescription: string;
  resumeText: string;
  companyName: string;
  role: string;
}

export interface CoverLetterResult {
  coverLetter: string;
}

export interface DashboardStats {
  totalApplications: number;
  interviews: number;
  offers: number;
  rejections: number;
  successRate: number;
  activeApplications: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  entityId: number;
  entityType: string;
  createdAt: string;
}

export interface MonthlyTrendItem {
  month: string;
  year: number;
  count: number;
}

export interface StatusBreakdownItem {
  status: string;
  count: number;
}

export type ListApplicationsParams = {
status?: string;
search?: string;
};

export type ListInterviewQuestionsParams = {
category?: string;
};