# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareerPilot AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List all job applications for the authenticated user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create a new job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get a specific application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Recruiters ──────────────────────────────────────────────────────────────
...
  /interview-questions:
    get:
      operationId: listInterviewQuestions
      tags: [interviewPrep]
      summary: List interview questions
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of interview questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewQuestion"
    post:
      operationId: createInterviewQuestion
      tags: [interviewPrep]
      summary: Create an interview question/note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewQuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewQuestion"
  /interview-questions/{id}:
    patch:
      operationId: updateInterviewQuestion
      tags: [interviewPrep]
      summary: Update an interview question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewQuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewQuestion"
        "404":
          description: Not found
    delete:
      operationId: deleteInterviewQuestion
      tags: [interviewPrep]
      summary: Delete an interview question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── AI Endpoints ─────────────────────────────────────────────────────────────
  /ai/analyze-resume:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: AI analysis of resume text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalysisResult"
  /ai/generate-cover-letter:
    post:
      operationId: generateCoverLetter
      tags: [applications]
      summary: Generate an AI cover letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterInput"
      responses:
        "200":
          description: Generated cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterResult"

  # ─── Dashboard / Analytics (wow endpoints) ────────────────────────────────────
...
components:
  schemas:
...
    Application:
      type: object
      required: [id, userId, company, role, status, appliedDate, createdAt, updatedAt]
      properties:
        id:          { type: integer }
        userId:      { type: string }
        company:     { type: string }
        role:        { type: string }
        location:    { type: ["string", "null"] }
        salary:      { type: ["string", "null"] }
        appliedDate: { type: string, format: date }
        status:
          type: string
          enum: [wishlist, applied, oa_received, interview_scheduled, final_round, offer, rejected]
        notes:       { type: ["string", "null"] }
        jobUrl:      { type: ["string", "null"] }
        createdAt:   { type: string, format: date-time }
        updatedAt:   { type: string, format: date-time }
    ApplicationInput:
      type: object
      required: [company, role, appliedDate, status]
      properties:
        company:     { type: string, minLength: 1 }
        role:        { type: string, minLength: 1 }
        location:    { type: string }
        salary:      { type: string }
        appliedDate: { type: string, format: date }
        status:
          type: string
          enum: [wishlist, applied, oa_received, interview_scheduled, final_round, offer, rejected]
        notes:       { type: string }
        jobUrl:      { type: string }
    ApplicationUpdate:
      type: object
      properties:
        company:     { type: string, minLength: 1 }
        role:        { type: string, minLength: 1 }
        location:    { type: string }
        salary:      { type: string }
        appliedDate: { type: string, format: date }
        status:
          type: string
          enum: [wishlist, applied, oa_received, interview_scheduled, final_round, offer, rejected]
        notes:       { type: string }
        jobUrl:      { type: string }

    # ── Recruiter ─────────────────────────────────────────────────────────────────
...
    InterviewQuestion:
      type: object
      required: [id, userId, question, category, createdAt, updatedAt]
      properties:
        id:        { type: integer }
        userId:    { type: string }
        question:  { type: string }
        answer:    { type: ["string", "null"] }
        category:
          type: string
          enum: [behavioral, technical, company_specific, general]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InterviewQuestionInput:
      type: object
      required: [question, category]
      properties:
        question:   { type: string, minLength: 1 }
        answer:     { type: string }
        category:
          type: string
          enum: [behavioral, technical, company_specific, general]
        difficulty:
          type: string
          enum: [easy, medium, hard]
    InterviewQuestionUpdate:
      type: object
      properties:
        question:   { type: string, minLength: 1 }
        answer:     { type: string }
        category:
          type: string
          enum: [behavioral, technical, company_specific, general]
        difficulty:
          type: string
          enum: [easy, medium, hard]

    # ── AI ─────────────────────────────────────────────────────────────────────────
    ResumeAnalysisInput:
      type: object
      required: [resumeText, jobDescription]
      properties:
        resumeText:     { type: string }
        jobDescription: { type: string }
    ResumeAnalysisResult:
      type: object
      required: [atsScore, feedback, missingKeywords, skillGaps, suggestions]
      properties:
        atsScore:        { type: integer }
        feedback:        { type: string }
        missingKeywords:
          type: array
          items: { type: string }
        skillGaps:
          type: array
          items: { type: string }
        suggestions:
          type: array
          items: { type: string }
    CoverLetterInput:
      type: object
      required: [jobDescription, resumeText, companyName, role]
      properties:
        jobDescription: { type: string }
        resumeText:     { type: string }
        companyName:    { type: string }
        role:           { type: string }
    CoverLetterResult:
      type: object
      required: [coverLetter]
      properties:
        coverLetter: { type: string }

    # ── Dashboard ──────────────────────────────────────────────────────────────────
...