# lib/api-zod/src/generated/types/coverLetterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetterInput {
  jobDescription: string;
  resumeText: string;
  companyName: string;
  role: string;
}