# lib/api-zod/src/generated/types/interviewQuestionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQuestionUpdateCategory } from './interviewQuestionUpdateCategory';
import type { InterviewQuestionUpdateDifficulty } from './interviewQuestionUpdateDifficulty';

export interface InterviewQuestionUpdate {
  /** @minLength 1 */
  question?: string;
  answer?: string;
  category?: InterviewQuestionUpdateCategory;
  difficulty?: InterviewQuestionUpdateDifficulty;
}