# lib/api-zod/src/generated/types/resumeAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeAnalysisResult {
  atsScore: number;
  feedback: string;
  missingKeywords: string[];
  skillGaps: string[];
  suggestions: string[];
}